Add identifier map to fgenesh taxonomy ouputs
