+++ webapp/utils/data/fgenesh/fetch_taxonomies/corrections.json
@@ -33,5 +33,16 @@
         "family": "Branchiostomidae",
         "order": "Amphioxiformes",
         "class": "Leptocardii"
+    },
+    "Nematocida parisii": {
+        "kingdom": "fungi",
+        "phylum": "Microsporidia",
+        "class": "incertae sedis",
+        "order": "incertae sedis",
+        "family": "Nosematidae"
+    },
+    "Nosema": {
+        "kingdom": "fungi",
+        "phylum": "Microsporidia"
     }
 }
