+++ webapp/utils/data/fgenesh/fetch_taxonomies/outputs/duplicates_entries.json
@@ -1,0 +1,173 @@
+{
+  "Dekkera bruxellensis_['yeast']": [
+    {
+      "name": "Dekkera bruxellensis",
+      "desc": "Brettanomyces bruxellensis",
+      "data": {
+        "usageKey": 2599345,
+        "scientificName": "Dekkera bruxellensis Van der Walt",
+        "canonicalName": "Dekkera bruxellensis",
+        "rank": "SPECIES",
+        "status": "ACCEPTED",
+        "confidence": 99,
+        "matchType": "EXACT",
+        "kingdom": "Fungi",
+        "phylum": "Ascomycota",
+        "order": "Saccharomycetales",
+        "family": "Pichiaceae",
+        "genus": "Dekkera",
+        "species": "Dekkera bruxellensis",
+        "kingdomKey": 5,
+        "phylumKey": 95,
+        "classKey": 182,
+        "orderKey": 1281,
+        "familyKey": 8404,
+        "genusKey": 2599343,
+        "speciesKey": 2599345,
+        "synonym": false,
+        "class": "Saccharomycetes"
+      }
+    }
+  ],
+  "Serendipita vermifera_['Serendipita indica, Achlamydospora indica']": [
+    {
+      "name": "Auriculariales",
+      "desc": "Sebacina vermifera",
+      "data": {
+        "usageKey": 2553903,
+        "scientificName": "Serendipita vermifera (Oberw.) P.Roberts",
+        "canonicalName": "Serendipita vermifera",
+        "rank": "SPECIES",
+        "status": "ACCEPTED",
+        "confidence": 99,
+        "matchType": "EXACT",
+        "kingdom": "Fungi",
+        "phylum": "Basidiomycota",
+        "order": "Auriculariales",
+        "genus": "Serendipita",
+        "species": "Serendipita vermifera",
+        "kingdomKey": 5,
+        "phylumKey": 34,
+        "classKey": 186,
+        "orderKey": 1062,
+        "genusKey": 7872390,
+        "speciesKey": 2553903,
+        "synonym": false,
+        "class": "Agaricomycetes"
+      }
+    }
+  ],
+  "Solanum lycopersicum_['tomato, Solanum lycopersicum']": [
+    {
+      "name": "Solanum lycopersicum",
+      "desc": "generic, tomato",
+      "data": {
+        "usageKey": 2930137,
+        "scientificName": "Solanum lycopersicum L.",
+        "canonicalName": "Solanum lycopersicum",
+        "rank": "SPECIES",
+        "status": "ACCEPTED",
+        "confidence": 98,
+        "matchType": "EXACT",
+        "kingdom": "Plantae",
+        "phylum": "Tracheophyta",
+        "order": "Solanales",
+        "family": "Solanaceae",
+        "genus": "Solanum",
+        "species": "Solanum lycopersicum",
+        "kingdomKey": 6,
+        "phylumKey": 7707728,
+        "classKey": 220,
+        "orderKey": 1176,
+        "familyKey": 7717,
+        "genusKey": 2928997,
+        "speciesKey": 2930137,
+        "synonym": false,
+        "class": "Magnoliopsida"
+      }
+    }
+  ],
+  "Strongylocentrotus purpuratus_['sea urchin']": [
+    {
+      "name": "Strongylocentrotus purpuratus",
+      "desc": "additional variant; sea urchin",
+      "data": {
+        "usageKey": 2278852,
+        "scientificName": "Strongylocentrotus purpuratus (Stimpson, 1857)",
+        "canonicalName": "Strongylocentrotus purpuratus",
+        "rank": "SPECIES",
+        "status": "ACCEPTED",
+        "confidence": 99,
+        "matchType": "EXACT",
+        "kingdom": "Animalia",
+        "phylum": "Echinodermata",
+        "order": "Camarodonta",
+        "family": "Strongylocentrotidae",
+        "genus": "Strongylocentrotus",
+        "species": "Strongylocentrotus purpuratus",
+        "kingdomKey": 1,
+        "phylumKey": 50,
+        "classKey": 221,
+        "orderKey": 5964707,
+        "familyKey": 9524,
+        "genusKey": 11308535,
+        "speciesKey": 2278852,
+        "synonym": false,
+        "class": "Echinoidea"
+      }
+    }
+  ],
+  "Rhizophydiales sp_['amphibian chytrid fungus']": [
+    {
+      "name": "Rhizophydiales",
+      "desc": "103 sequences, chytrids",
+      "data": {
+        "usageKey": 1117,
+        "scientificName": "Rhizophydiales",
+        "canonicalName": "Rhizophydiales",
+        "rank": "ORDER",
+        "status": "ACCEPTED",
+        "confidence": 94,
+        "matchType": "HIGHERRANK",
+        "kingdom": "Fungi",
+        "phylum": "Chytridiomycota",
+        "order": "Rhizophydiales",
+        "kingdomKey": 5,
+        "phylumKey": 94,
+        "classKey": 10689325,
+        "orderKey": 1117,
+        "synonym": false,
+        "class": "Rhizophydiomycetes"
+      }
+    }
+  ],
+  "Danio rerio_['Zebrafish, additional variant']": [
+    {
+      "name": "Danio rerio",
+      "desc": "Zebrafish, leopard danio",
+      "data": {
+        "usageKey": 9797255,
+        "scientificName": "Danio rerio (Hamilton, 1822)",
+        "canonicalName": "Danio rerio",
+        "rank": "SPECIES",
+        "status": "ACCEPTED",
+        "confidence": 97,
+        "matchType": "EXACT",
+        "kingdom": "Animalia",
+        "phylum": "Chordata",
+        "order": "Cypriniformes",
+        "family": "Cyprinidae",
+        "genus": "Danio",
+        "species": "Danio rerio",
+        "kingdomKey": 1,
+        "phylumKey": 44,
+        "orderKey": 1153,
+        "familyKey": 7336,
+        "genusKey": 2362675,
+        "speciesKey": 9797255,
+        "synonym": false,
+        "class": "Actinopterygii"
+      }
+    }
+  ]
+}
